Major update on concurrency
	transmitters loose voltage, each at different rates (arbetrary for now)
	synapses grow only if in a pool of unclaimed transmitters
	dendrites grow until transmitters are found

+++ src/main/java/com/trifidearth/zulu/coordinate/CoordinatePair.java
@@ -54,7 +54,7 @@ public class CoordinatePair {
         growing = growPoint;
     }
     
-    public void growRandom(int growDistance, CoordinateBounds bounds) {
+    public void growRandom(double growDistance, CoordinateBounds bounds) {
         Coordinate growPoint;
         do {
             growPoint = calculate(fixed,
